Fixes from remarks 2.1
