+++ aleksandrov.igor/P2/main.cpp
@@ -1,36 +1,9 @@
-#include "taylor.h"
-#include "row.h"
 #include <iostream>
 #include <stdexcept>
-#include <cmath>
 #include <iomanip>
 #include <cstring>
-
-void aleksandrov::printRow(double i, size_t k, double error)
-{
-  constexpr const char* errorMessage = "<MATH ERROR>";
-  constexpr size_t columnWidth = std::strlen(errorMessage);
-  std::cout << std::setw(columnWidth) << std::left << std::fixed << std::setprecision(5) << i;
-  try
-  {
-    std::cout << std::setw(columnWidth) << std::left << std::fixed << std::setprecision(5);
-    std::cout << taylor(i, k, error);
-  }
-  catch (const std::logic_error& e)
-  {
-    std::cout << std::setw(columnWidth) << std::left;
-    std::cout << errorMessage;
-  }
-  try
-  {
-    std::cout << std::setw(columnWidth) << std::right << std::fixed << std::setprecision(5);
-    std::cout << sinxDivx(i);
-  }
-  catch (const std::logic_error& e)
-  {
-    std::cout << std::setw(columnWidth) << std::left << errorMessage;
-  }
-}
+#include "taylor.h"
+#include "row.h"
 
 int main()
 {
